Fixed yellow box warning.
Fixed yellow box warning caused by putting FlatList inside ScrollView.
