+++ components/shared/StepSelectionTab.js
@@ -1,0 +1,3 @@
+import React from "react";
+
+export default class StepSelectionTab extends React.Component {}
