+++ resources/StepOptions.js
@@ -1,0 +1,6 @@
+export default stepOptions = [
+  {
+    id: "0",
+    name: "Pour Water",
+  },
+];
