+++ screens/recipes/CreateRecipeScreen.js
@@ -1,5 +1,5 @@
 import React from "react";
-import CreateRecipe from "../../components/CreateRecipe";
+import CreateRecipe from "../../components/recipe/create/CreateRecipe";
 
 export default class CreateRecipeScreen extends React.Component {
   render() {
